+++ docs/source/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `moldrug.constraintconf.gen_aligned_conf`
 - In case that `moldrug.constraintconf.generate_conformers` fails with `rdkit.Chem.AllChem.ConstrainedEmbed` it will try with `moldrug.constraintconf.gen_aligned_conf`.
 
+### Changed
+
+- The output error name when constraint fails has a idx prefix. E.g. 33_conf_27_error.pbz2 now is: idx_33_conf_27_error.pbz2. Now it is easy to delete all of this files at the end of the simulation if they are not needed.
+
 ## [3.0.3] - 2022.11.26
 
 ### Added